+++ src/main/java/com/bulletphysics/collision/broadphase/AxisSweep3.java
@@ -26,13 +26,13 @@
 
 package com.bulletphysics.collision.broadphase;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 /**
  * AxisSweep3 is an efficient implementation of the 3D axis sweep and prune broadphase.<p>
  * 
  * It uses arrays rather then lists for storage of the 3 axis. Also it operates using 16 bit
- * integer coordinates instead of floats. For large worlds and many objects, use {@link AxisSweep3_32}
+ * integer coordinates instead of doubles. For large worlds and many objects, use {@link AxisSweep3_32}
  * instead. AxisSweep3_32 has higher precision and allows more than 16384 objects at the cost
  * of more memory and bit of performance.
  *
@@ -40,15 +40,15 @@ import javax.vecmath.Vector3f;
  */
 public class AxisSweep3 extends AxisSweep3Internal {
 
-	public AxisSweep3(Vector3f worldAabbMin, Vector3f worldAabbMax) {
+	public AxisSweep3(Vector3d worldAabbMin, Vector3d worldAabbMax) {
 		this(worldAabbMin, worldAabbMax, 16384, null);
 	}
 
-	public AxisSweep3(Vector3f worldAabbMin, Vector3f worldAabbMax, int maxHandles) {
+	public AxisSweep3(Vector3d worldAabbMin, Vector3d worldAabbMax, int maxHandles) {
 		this(worldAabbMin, worldAabbMax, maxHandles, null);
 	}
 	
-	public AxisSweep3(Vector3f worldAabbMin, Vector3f worldAabbMax, int maxHandles/* = 16384*/, OverlappingPairCache pairCache/* = 0*/) {
+	public AxisSweep3(Vector3d worldAabbMin, Vector3d worldAabbMax, int maxHandles/* = 16384*/, OverlappingPairCache pairCache/* = 0*/) {
 		super(worldAabbMin, worldAabbMax, 0xfffe, 0xffff, maxHandles, pairCache);
 		// 1 handle is reserved as sentinel
 		assert (maxHandles > 1 && maxHandles < 32767);
