+++ src/main/java/com/bulletphysics/collision/broadphase/BroadphaseInterface.java
@@ -23,7 +23,7 @@
 
 package com.bulletphysics.collision.broadphase;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 /**
  * BroadphaseInterface for AABB overlapping object pairs.
@@ -32,11 +32,11 @@ import javax.vecmath.Vector3f;
  */
 public abstract class BroadphaseInterface {
 
-	public abstract BroadphaseProxy createProxy(Vector3f aabbMin, Vector3f aabbMax, BroadphaseNativeType shapeType, Object userPtr, short collisionFilterGroup, short collisionFilterMask, Dispatcher dispatcher, Object multiSapProxy);
+	public abstract BroadphaseProxy createProxy(Vector3d aabbMin, Vector3d aabbMax, BroadphaseNativeType shapeType, Object userPtr, short collisionFilterGroup, short collisionFilterMask, Dispatcher dispatcher, Object multiSapProxy);
 
 	public abstract void destroyProxy(BroadphaseProxy proxy, Dispatcher dispatcher);
 
-	public abstract void setAabb(BroadphaseProxy proxy, Vector3f aabbMin, Vector3f aabbMax, Dispatcher dispatcher);
+	public abstract void setAabb(BroadphaseProxy proxy, Vector3d aabbMin, Vector3d aabbMax, Dispatcher dispatcher);
 
 	///calculateOverlappingPairs is optional: incremental algorithms (sweep and prune) might do it during the set aabb
 	public abstract void calculateOverlappingPairs(Dispatcher dispatcher);
@@ -45,7 +45,7 @@ public abstract class BroadphaseInterface {
 	
 	///getAabb returns the axis aligned bounding box in the 'global' coordinate frame
 	///will add some transform later
-	public abstract void getBroadphaseAabb(Vector3f aabbMin, Vector3f aabbMax);
+	public abstract void getBroadphaseAabb(Vector3d aabbMin, Vector3d aabbMax);
 
 	public abstract void printStats();
